pyproject.toml
[project]
name = "xpiano"
version = "0.1.0"
description = "CLI-first AI piano coach MVP"
readme = "docs/prd.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.39",
    "fastdtw>=0.3",
    "jsonschema>=4.20",
    "mido>=1.3",
    "numpy>=1.24",
    "pretty_midi>=0.2",
    "python-rtmidi>=1.5",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.scripts]
xpiano = "xpiano.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
